+++ src/maite_datasets/wrappers/__init__.py
@@ -3,6 +3,6 @@ import importlib.util
 __all__ = []
 
 if importlib.util.find_spec("torch") is not None and importlib.util.find_spec("torchvision") is not None:
-    from ._torch import TorchWrapper
+    from ._torch import TorchvisionWrapper
 
-    __all__ += ["TorchWrapper"]
+    __all__ += ["TorchvisionWrapper"]